requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
openai-whisper>=20230314